Corrección de detalles. Se ha creado las rutas para request y request_type, así como también se han levantado en e servidor

+++ src/models/requests.model.ts
@@ -18,7 +18,7 @@ const RequestsModel = {
         },*/
     },
     amount: {
-        type: DataTypes.DOUBLE,
+        type: DataTypes.DECIMAL(12,2),
         allowNull: false,
     },
     request_status: {
